Try to get HTTP 304 from upstream server
Keep track of the ETag returned from upstream server and mention it
in If-None-Match whenever refreshing the cache. This allows the upstream
server to return a 304 without body instead of a 200 with the file we
already have.

Notice RFC2616 section 13.3.4 says if we are going to try to play it
smart we MUST use ETag and we SHOULD use Last-Modified. So, let's keep
it simple and stick to just ETag for now.